RLP + basic smart contract (#16)
* marking testcases as invalid
* initial script hashing
* Adjust to other script types
* generate correct expected data
* extend test expected results calculation
* complete basic formatting
* basic transaction reprensentation
* bump version number
* remove obsolete test

+++ app/src/parser.h
@@ -38,7 +38,7 @@ parser_error_t parser_getNumItems(const parser_context_t *ctx, uint8_t *num_item
 
 // retrieves a readable output for each field / page
 parser_error_t parser_getItem(const parser_context_t *ctx,
-                              uint8_t displayIdx,
+                              uint16_t displayIdx,
                               char *outKey, uint16_t outKeyLen,
                               char *outVal, uint16_t outValLen,
                               uint8_t pageIdx, uint8_t *pageCount);
